Moves sandbox main file to manual testing/experimentation folder
