+++ src/system/Serial.cpp
@@ -4,6 +4,7 @@
 #include <string.h>
 #include <stdexcept>
 
+#include <fcntl.h> // Contains file controls like O_RDWR
 #include <errno.h> // Error integer and strerror() function
 #include <termios.h> // Contains POSIX terminal control definitions
 #include <unistd.h> // write(), read(), close()
